Refactoring and moving functions into multiple script files
