+++ assets/displayStats.js
@@ -1,9 +1,7 @@
-export function displayStats(statsText, seconds, userMistakesCount, userKeyTypeCount) {
+export function displayStats(userMistakesCount, userKeyTypeCount) {
   if (userMistakesCount == 0 && userKeyTypeCount == 0) {
-    statsText.innerHTML = 'Seconds: ' + seconds + ', ' + '0' + '% Wrong Chars';
-
-    return;
+    return '0' + '% Wrong Chars';
   }
 
-  statsText.innerHTML = 'Seconds: ' + seconds + ', ' + Math.round((userMistakesCount * 100 / userKeyTypeCount * 100) / 100) + '% Wrong Chars';
+  return  Math.round((userMistakesCount * 100 / userKeyTypeCount * 100) / 100) + '% Wrong Chars';
 }
