--- assets/timeHandler.js
@@ -1,5 +1,0 @@
-export function incrementSeconds() {
-  intervalIsUsed = true;
-  seconds += 1;
-  displayStats(statsText, seconds, userMistakesCount, userKeyTypeCount);
-}
