+++ index.html
@@ -29,5 +29,5 @@
     </div>
     </div>
   </body>
-  <script type="module" src="assets/getWords.js"></script>
+  <script type="module" src="assets/main.js"></script>
 </html>
